+ add starcoder inference test with tis model https://huggingface.co/mike-ravkine/gpt_bigcode-santacoder-GGML + update ggml to d2b178ee747cfb0a0a1892da362d068080659170 + update llama to 6e7cca404748dd4b1a3affd0d1296e37f4ac0a6f

+++ LLMFarm_core/LLaMa.swift
@@ -39,7 +39,7 @@ public class LLaMa: GPTBase {
         var params = llama_context_default_params()
         params.n_ctx = contextParams.context
 //        params.n_parts = contextParams.parts
-        params.seed = contextParams.seed
+        params.seed = UInt32(contextParams.seed)
         params.f16_kv = contextParams.f16Kv
         params.logits_all = contextParams.logitsAll
         params.vocab_only = contextParams.vocabOnly
